File and Folder names in the renamer should not be encoded to their locale encoding. They should remain in UTF-8.
Encoding from UTF-8 to the locale encoding could have undesired effects because python naively decodes strings to UTF-8 during os calls. If the locale encoding is not UTF-8 or a subset of UTF-8, errors will occur when characters unique to the locale encoding are present.

For example: on Windows, the locale encoding is 'cp1252' and it is not a subset of UTF-8.
The EN DASH character in UTF-8 is /u2013
The EN DASH character in cp1252 is /x96

If we were to encode a string containing the EN DASH character to the local encoding, a /x96 byte would then be present.
When Python makes its os call for creating a directory or moving/copying a file, it would naively decode the path back to UTF-8. It then would think that the /x96 character was a UTF-8 control character.